requirements-dev.txt
# ===========================================
# Condominios Manager - Development Requirements
# ===========================================
# Compatible with Python 3.12, 3.13, and 3.14
# Install: uv pip install -r requirements-dev.txt

# Include production dependencies
-r requirements.txt

# ===========================================
# Linting & Formatting (Ruff replaces black, isort, flake8, pylint)
# ===========================================
ruff>=0.11.0,<1.0

# ===========================================
# Type Checking
# ===========================================
mypy>=1.16.0,<2.0
django-stubs[compatible-mypy]>=5.2.0,<6.0
djangorestframework-stubs[compatible-mypy]>=3.15.0,<4.0
types-python-dateutil>=2.9.0,<3.0
types-requests>=2.31.0,<3.0

# ===========================================
# Security
# ===========================================
bandit>=1.9.0,<2.0
safety>=3.0.0,<4.0

# ===========================================
# Testing
# ===========================================
pytest>=9.0.0,<10.0
pytest-django>=4.10.0,<5.0
pytest-cov>=6.0.0,<8.0
pytest-mock>=3.14.0,<4.0
pytest-xdist>=3.8.0,<4.0
pytest-timeout>=2.3.0,<3.0
pytest-env>=1.1.0,<2.0
pytest-asyncio>=1.0.0,<2.0
factory-boy>=3.3.0,<4.0
faker>=29.0.0,<41.0
freezegun>=1.4.0,<2.0
model-bakery>=1.20.0,<2.0
responses>=0.25.0,<1.0

# Test Coverage
coverage>=7.6.0,<8.0

# ===========================================
# Pre-commit Hooks
# ===========================================
pre-commit>=4.0.0,<5.0

# ===========================================
# Debugging
# ===========================================
ipython>=9.0.0,<10.0
ipdb>=0.13.0,<1.0
django-debug-toolbar>=5.0.0,<7.0
